feat: add per-pixel z occlusion and z shader
PixelGrid now has 2 Vecs: pixel colors and pixels_zs. Accessing a pixel with the get_pixel() function would return a mutable reference to the pixel's z and the pixel's color.

Also created crates/examples/shaders directory which in the future would showcase the shader capabilities of the engine. Added one shader, the z_shader that just shows the z values of the pixel.