refactor: create (OCS) object, component, system system
Something in between what unity does (objects with components) but with
the additions of systems, which manage objects in large quantities.

This is a huge commit as almost all systems had to be changed including
the mesh rendering one. Different rendering systems are now opt-in.

Vertex has been moved to math. The struct Object is now a generic object
with the ability to have components. And objects that can be rendered
as a mesh are Object with MeshRenderable and Transform components.
Some dependencies required for downcasting have also been added
(downcast_rs)