fix: do not account for rotation when calculating mesh center, only account for position.
Additionally, account for overall mesh position and rotation on a per vertex basis. In the mesh2mesh collision detection algorithm when raycasting from the center towards the vertices. Previously, the center was calculated considering shape position and rotation, but the vertex weren't. Now the center is calculated only using the position and the vertices with both the position and rotation.